:recycle: Trukket dekorator-kode ut av alle templates
