+++ @navikt/core/react/templates/404/med-cta.stories.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import React from "react";
 import {
   BodyShort,
   Box,
@@ -9,6 +9,7 @@ import {
   Page,
   VStack,
 } from "@navikt/ds-react";
+import { Env, Footer, Header, useDekorator } from "../DemoDecorator";
 
 export default {
   title: "templates/404/Med CTA",
@@ -51,49 +52,6 @@ const Example = () => {
     </Page>
   );
 };
-
-function Header() {
-  return <div id="decorator-header" />;
-}
-
-function Footer() {
-  return <div id="decorator-footer" />;
-}
-
-const MILJO_URL = "https://www.nav.no/dekoratoren";
-
-function Env() {
-  return (
-    <div
-      id="decorator-env"
-      data-src={`${MILJO_URL}/env?context=privatperson`}
-    />
-  );
-}
-
-/**
- * OBS: Dette er ikke anbefalt metode for å laste dekoratør!
- * Se `nav-dekoratoren`-dokumentasjon for riktig implementasjon
- * @see https://github.com/navikt/nav-dekoratoren
- */
-function useDekorator() {
-  useEffect(() => {
-    const script = document.createElement("script");
-    script.src = `${MILJO_URL}/client.js`;
-    script.async = true;
-    document.body.appendChild(script);
-
-    const styles = document.createElement("link");
-    styles.href = `${MILJO_URL}/css/client.css`;
-    styles.rel = "stylesheet";
-    document.head.appendChild(styles);
-
-    return () => {
-      document.body.removeChild(script);
-      document.head.removeChild(styles);
-    };
-  }, []);
-}
 
 export const Demo = {
   render: Example,
